Chore/final refactor (#428)
* Introduced BackgroundScript

* Added controller base

* Made Controller modules to inherit from ControllerBase

* Fixed lint errors

* Fixed linter config
